# openapi.json
{"openapi":"3.1.0","info":{"title":"Payment Middleware","description":"Middleware translating ISO 8583 messages to ILP/Rafiki for inter-coop transfers","version":"0.2.0"},"paths":{"/health":{"get":{"tags":["health"],"summary":"Health Check","description":"Health check endpoint used by developers, Docker, and orchestration tools.","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/wallets":{"get":{"tags":["wallets"],"summary":"List Wallet Mappings","description":"List all cooperative to wallet URL mappings.","operationId":"list_wallet_mappings_wallets_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/transactions":{"get":{"tags":["transactions"],"summary":"List Transactions","description":"List all transactions with optional filtering.","operationId":"list_transactions_transactions_get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":100,"title":"Limit"}},{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/transactions/{transaction_id}":{"get":{"tags":["transactions"],"summary":"Get Transaction","description":"Get transaction details by ID.","operationId":"get_transaction_transactions__transaction_id__get","parameters":[{"name":"transaction_id","in":"path","required":true,"schema":{"type":"integer","title":"Transaction Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/iso8583/process":{"post":{"tags":["iso8583"],"summary":"Process Iso8583 Message","description":"Process an ISO 8583 financial request and execute transfer via Rafiki.\n\nArgs:\n    source_account: Source account number (ISO Field 2)\n    amount: Transfer amount\n    currency: Currency code (ISO 4217, default USD=840)\n    source_coop_id: Source cooperative ID (ISO Field 32)\n    dest_coop_wallet: Destination coop's Rafiki wallet URL (overrides dest_coop_id)\n    dest_coop_id: Destination cooperative ID (resolved via mapping)\n    use_mock: Use mock Rafiki client for testing\n    rafiki_url: Rafiki server URL for real transfers\n\nReturns:\n    ISO 8583 response message","operationId":"process_iso8583_message_iso8583_process_post","parameters":[{"name":"source_account","in":"query","required":true,"schema":{"type":"string","title":"Source Account"}},{"name":"amount","in":"query","required":true,"schema":{"type":"number","title":"Amount"}},{"name":"currency","in":"query","required":false,"schema":{"type":"string","default":"840","title":"Currency"}},{"name":"source_coop_id","in":"query","required":false,"schema":{"type":"string","default":"","title":"Source Coop Id"}},{"name":"dest_coop_wallet","in":"query","required":false,"schema":{"type":"string","default":"","title":"Dest Coop Wallet"}},{"name":"dest_coop_id","in":"query","required":false,"schema":{"type":"string","default":"","title":"Dest Coop Id"}},{"name":"use_mock","in":"query","required":false,"schema":{"type":"boolean","default":true,"title":"Use Mock"}},{"name":"rafiki_url","in":"query","required":false,"schema":{"type":"string","default":"http://localhost:3000","title":"Rafiki Url"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/transfer":{"post":{"tags":["transfer"],"summary":"Initiate Transfer","description":"Simplified endpoint to initiate transfer between coops.\nThis is a higher-level API that handles ISO 8583 internally.","operationId":"initiate_transfer_transfer_post","parameters":[{"name":"source_account","in":"query","required":true,"schema":{"type":"string","title":"Source Account"}},{"name":"amount","in":"query","required":true,"schema":{"type":"number","title":"Amount"}},{"name":"dest_coop_wallet","in":"query","required":true,"schema":{"type":"string","title":"Dest Coop Wallet"}},{"name":"currency","in":"query","required":false,"schema":{"type":"string","default":"840","title":"Currency"}},{"name":"source_coop_id","in":"query","required":false,"schema":{"type":"string","default":"","title":"Source Coop Id"}},{"name":"use_mock","in":"query","required":false,"schema":{"type":"boolean","default":true,"title":"Use Mock"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/iso8583/parse":{"post":{"tags":["iso8583"],"summary":"Parse Iso8583 Hex","description":"Parse raw ISO 8583 message from hex string using pyiso8583.\nUseful for debugging incoming messages.\nProvide hex_data as query parameter or in request body as JSON: {\"hex_data\": \"...\"}","operationId":"parse_iso8583_hex_iso8583_parse_post","parameters":[{"name":"hex_data","in":"query","required":false,"schema":{"type":"string","title":"Hex Data"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","title":"Body"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}